implemented changes acording to PR comments
