+++ src/App.tsx
@@ -1,12 +1,14 @@
 import React from 'react'
 import Header from './components/Header'
 import Events from './components/Events'
+import AddEventButton from './components/AddEventButton/AddEventButton'
 
 const App = () => {
     return (
         <>
             <Header />
             <Events />
+            <AddEventButton />
         </>
     )
 }
